alteracoes

+++ controlelista/Cliente.java
@@ -6,9 +6,9 @@ public class Cliente {
     private long cpf;
     private int numeroCarteiraMotorista;
     private String endereco;
-    private int telefone;
+    private long telefone;
 
-    public Cliente(long cpf, String nome, int numeroCarteiraMotorista, String endereco, int telefone){
+    public Cliente(long cpf, String nome, int numeroCarteiraMotorista, String endereco, long telefone){
         this.nome = nome;
         this.cpf = cpf;
         this.numeroCarteiraMotorista = numeroCarteiraMotorista;
@@ -28,19 +28,19 @@ public class Cliente {
 
     public String getEndereco() { return endereco; }
 
-    public int getTelefone() { return telefone; }
+    public long getTelefone() { return telefone; }
 
     public String getResumo() {
         return "Nome: " + nome + " CPF: " + cpf;
     }
 
-    public void setNome(String nome) { this.nome=nome; }
+    public void setNome(String nome) { this.nome = nome; }
 
     public void setNumeroCarteira(int numeroCarteiraMotorista) { this.numeroCarteiraMotorista = numeroCarteiraMotorista; }
 
     public void setEndereco(String endereco) { this.endereco = endereco; }
 
-    public void setTelefone(int telefone) { this.telefone = telefone; }
+    public void setTelefone(long telefone) { this.telefone = telefone; }
 
     public String toString() {
         return "Nome: " + nome +
